+++ tests/deployment/aws_ec2/test_ec2_deployment_operator.py
@@ -1,6 +1,5 @@
 import os
 
-import boto3
 from mock import MagicMock, patch
 
 from bentoml.yatai.proto import status_pb2
@@ -16,7 +15,6 @@ from bentoml.yatai.proto.repository_pb2 import (
     GetBentoResponse,
 )
 from bentoml.yatai.deployment.aws_ec2.operator import (
-    _get_ecr_password,
     _make_cloudformation_template,
     AwsEc2DeploymentOperator,
 )
@@ -68,28 +66,6 @@ def generate_ec2_deployment_pb():
     test_deployment_pb.spec.aws_ec2_operator_config.autoscale_max_size = 1
 
     return test_deployment_pb
-
-
-def test_get_ecr_password():
-    def mock_ecr_client_auth_token(
-        self, op_name, args
-    ):  # pylint: disable=unused-argument
-        if op_name == "GetAuthorizationToken":
-            return {
-                "authorizationData": [
-                    {
-                        "authorizationToken": mock_registry_auth_token,
-                        "proxyEndpoint": mock_registry_endpoint,
-                    }
-                ]
-            }
-
-    with patch(
-        "botocore.client.BaseClient._make_api_call", new=mock_ecr_client_auth_token
-    ):
-        token, ep = _get_ecr_password(mock_registry_id, mock_region)
-        assert token == mock_registry_auth_token
-        assert ep == mock_registry_endpoint
 
 
 def test_make_cloudformation_template(tmpdir):
